+++ docs/make.jl
@@ -7,6 +7,7 @@ makedocs(;
     pages = [
         "Home" => "index.md",
         "Data Validation" => "data-validation.md",
+        "Bayesian Fitting" => "fitting.md",
         "API" => "api.md",
     ],
     warnonly = [:missing_docs],